requirement.txt
aiohttp==3.7.4.post0
amqp==5.1.1
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.3
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asn1crypto==1.4.0
asttokens==2.2.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==20.8b1
bleach==4.1.0
boto3==1.16.58
botocore==1.19.58
Brotli==1.0.9
bson==0.5.10
celery==5.2.7
certifi==2020.6.20
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
crypto==1.4.1
cryptography==3.4.6
cssselect2==0.7.0
curlify==2.2.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
device-detector==5.0.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==0.5.0
Django==4.1.7
django-autoslug-iplweb==1.9.4.dev0
django-braces==1.14.0
django-celery-beat==2.2.0
django-celery-results==2.2.0
django-cities-light==3.5.0
django-ckeditor==6.0.0
django-core==1.4.1
django-cors-headers==3.13.0
django-crispy-forms==1.11.2
django-debug-toolbar==1.11
django-defender==0.8.0
django-extensions==3.1.0
django-filter==22.1
django-import-export==3.1.0
django-ipware==3.0.2
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-oauth-toolkit==1.3.3
django-oauth2==3.2
django-ratelimit==3.0.1
django-rename-app==0.1.3
django-reset-migrations==0.4.0
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==0.9.7
django-s3-storage==0.13.4
django-simple-history==2.12.0
django-smtp-ssl==1.0
django-storages==1.11.1
django-structlog==2.1.0
django-su==0.8.0
django-tempus-dominus==5.1.2.13
django-timezone-field==4.1.1
django-tracking==0.2.7
django-truncate==0.1
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.15.2
drf-extensions==0.4.0
drf-yasg==1.15.0
endesive==2.0.3
entrypoints==0.3
et-xmlfile==1.1.0
eventlet==0.24.1
executing==1.2.0
facebook-business==11.0.0
filelock==3.10.3
fonttools==4.38.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
html2text==2020.1.16
html5lib==1.1
identify==2.5.21
idna==2.10
importlib-metadata==6.1.0
importlib-resources==3.3.0
inflection==0.5.1
ipdb==0.13.13
ipykernel==6.7.0
ipython==8.11.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.0.1
jmespath==0.10.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jwcrypto==1.0
kombu==5.2.4
listener==2.0.0a1
lxml==4.6.2
mailchimp3==3.0.14
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
Minidetector==0.1
mistune==0.8.4
monotonic==1.6
multidict==5.1.0
mypy-extensions==0.4.3
Naked==0.1.31
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
nodeenv==1.7.0
notebook==6.4.7
num2words==0.5.10
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.1
oscrypto==1.2.1
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.7.2
parso==0.8.3
pathspec==0.8.1
pdfkit==1.0.0
pendulum==2.0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pre-commit==3.2.0
pre-commit-hooks==4.1.0
progressbar2==3.53.2
prometheus-client==0.12.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycountry==20.7.3
pycparser==2.21
pycryptodome==3.10.1
pycryptodomex==3.10.1
pydash==4.7.4
pydyf==0.5.0
Pygments==2.14.0
PyJWT==2.4.0
PyKCS11==1.5.10
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==3.0.7
PyPDF2==1.26.0
pyphen==0.13.2
pyrsistent==0.18.1
PySocks==1.7.1
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.1
python-magic==0.4.15
python-memcached==1.59
python-utils==2.5.6
python3-openid==3.2.0
pytube==10.5.3
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
redis==3.5.3
regex==2021.4.4
reportlab==3.5.67
requests==2.24.0
requests-oauthlib==1.3.0
rest-framework-generic-relations==1.2.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.3.4
Send2Trash==1.8.0
shellescape==3.8.1
shortuuid==1.0.1
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
SQLAlchemy==1.4.40
sqlparse==0.4.3
stack-data==0.6.2
stripe==2.10.1
structlog==20.1.0
tablib==3.3.0
terminado==0.12.1
testpath==0.5.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.9.0
twilio==6.24.1
typed-ast==1.4.3
typing_extensions==4.5.0
tzdata==2022.1
Unidecode==1.3.0
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.6
weasyprint==58.0
webencodings==0.5.1
websocket-client==1.2.3
whitenoise==3.3.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zipp==3.15.0
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.2.2
